+++ jlxd-core/src/main/java/au/com/jcloud/lxd/util/FormatUtils.java
@@ -25,11 +25,11 @@ public class FormatUtils {
 		return DATE_FORMAT_ISO8601_SHORT_EXT.format(date);
 	}
 
-	public static String convertIntegerToMB(Integer input) {
+	public static String convertIntegerToMB(Long input) {
 		return String.valueOf(new BigDecimal(Double.valueOf(input) / 1024 / 1024).setScale(2, RoundingMode.HALF_UP));
 	}
 	
-	public static String convertIntegerToKB(Integer input) {
+	public static String convertIntegerToKB(Long input) {
 		return String.valueOf(new BigDecimal(Double.valueOf(input) / 1024).setScale(2, RoundingMode.HALF_UP));
 	}
